disable protected-mode in sentinel
redis 3.2 introduced so-called `protected mode`. it automatically kicks
in when certain conditions are met. but sentinel, which is implemented
in redis itself, does not support `AUTH`. even if you _protect_ the
redis, you cannot protect sentinel as it does not support authentication
at all.

The author states in https://github.com/antirez/redis/pull/3329 that
sentinel does not support AUTH.